Don't grab disabled domains when generating configuration fixes #272
Currently when we disable a sub-domain, the DNS entry gets removed, if
dns is locally managed, but the apache vhost is always added to
vhosts_all.conf.

This means that for domains that don't have local DNS management, the
disable link has no effect!

We'll place the check for disabled domains before the check for the type
of entry: checking for simple equality on a column is simpler and faster
than transforming values from two columns with tolower(), so it's a fast
way to reduce the number of rows that should be evaluated.

Since I'm changing an SQL request that has trailing spaces, I'm removing
them at the same time. They are unneeded and undesirable.